Moved helpers (#147)
