--- src/mlmodule/v2/helpers/types.py
@@ -1,6 +1,0 @@
-from typing import Union
-
-import numpy as np
-import torch
-
-NumericArrayTypes = Union[torch.Tensor, np.ndarray]
