+++ tests/models/test_vinvl.py
@@ -1,13 +1,13 @@
 import os
 
 from mlmodule.callbacks.memory import CollectBoundingBoxesInMemory
+from mlmodule.helpers.files import list_files_in_dir
 from mlmodule.models.vinvl.modules import TorchVinVLDetectorModule
 from mlmodule.states import StateKey
 from mlmodule.stores import Store
 from mlmodule.torch.datasets import ImageDataset, LocalBinaryFilesDataset
 from mlmodule.torch.options import TorchRunnerOptions
 from mlmodule.torch.runners import TorchInferenceRunner
-from mlmodule.utils import list_files_in_dir
 
 
 def test_vinvl_object_detection(torch_device):
