+++ tests/stores/test_provider_stores.py
@@ -5,7 +5,7 @@ import pytest
 from mlmodule.models.types import ModelWithState
 from mlmodule.states import StateKey
 from mlmodule.stores import Store
-from mlmodule.v2.testing import ModuleTestConfiguration
+from mlmodule.testing import ModuleTestConfiguration
 
 
 @pytest.mark.slow
